LibNeuroN/PythonManager/Extra/requirements.txt
voicevox-client
speechrecognition
asyncio